+++ num_enum/tests/try_build/compile_fail/conflicting_derive.rs
@@ -1,0 +1,11 @@
+#[derive(num_enum::FromPrimitive, num_enum::TryFromPrimitive)]
+#[repr(u8)]
+enum Numbers {
+    Zero,
+    #[num_enum(default)]
+    One,
+}
+
+fn main() {
+
+}
