+++ num_enum/tests/try_build/compile_fail/missing_repr.rs
