+++ num_enum/tests/try_build/compile_fail/repr_c.stderr
