+++ renamed_num_enum/src/lib.rs
@@ -1,8 +1,10 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 extern crate alloc;
+
 use ::alloc::format;
 use ::core::convert::TryFrom;
+
 use ::renamed::{IntoPrimitive, TryFromPrimitive, UnsafeFromPrimitive};
 
 #[derive(Debug, PartialEq, Eq, IntoPrimitive, TryFromPrimitive, UnsafeFromPrimitive)]
